+++ src/main/kotlin/lotto/model/LottoPrice.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.model
 
 import lotto.Const
