+++ src/main/kotlin/lotto/process/LottoPurchase.kt
@@ -1,6 +1,8 @@
-package lotto.domain
+package lotto.process
 
 import lotto.Const
+import lotto.model.LottoPrice
+import lotto.model.LottoTicket
 import java.util.Collections
 
 class LottoPurchase {
