+++ src/test/kotlin/lotto/process/LottoScoreTest.kt
@@ -1,5 +1,9 @@
-package lotto.domain
+package lotto.process
 
+import lotto.model.LottoNumber
+import lotto.model.LottoPrice
+import lotto.model.LottoPrize
+import lotto.view.LottoResult
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
@@ -58,5 +62,5 @@ internal class LottoScoreTest {
     }
 
     private fun LottoTicket(vararg numbers: Int) =
-        LottoTicket(numbers.map { LottoNumber(it) }.toSet())
+        lotto.model.LottoTicket(numbers.map { LottoNumber(it) }.toSet())
 }
